refactor: getPopularBooks 위치를 app/_services/books로 이동
