+++ src/app/(home)/_services/books.ts
@@ -1,21 +1,6 @@
-import { fetchPopularBooks, fetchRisingBooks } from '@/services';
-import { GetPopularBooksParamsProps } from '@/services/endpoints/naruEndpoint';
+import { fetchRisingBooks } from '@/services';
 import { NaruApiBookData } from '@/types';
 import { formatNaruApiBookDataToBookItemData } from '@/utils';
-
-export const getPopularBooks = async (props: GetPopularBooksParamsProps) => {
-  const data = await fetchPopularBooks(props);
-  const { docs, numFound, resultNum } = data.response;
-  const books = docs.map(({ doc }: { doc: NaruApiBookData }) => {
-    return formatNaruApiBookDataToBookItemData(doc);
-  });
-
-  if (!props.pageNo) return { books, isLastPage: true };
-
-  const isLastPage = (Number(props.pageNo) - 1) * Number(props.pageSize) + resultNum >= numFound;
-
-  return { books, isLastPage };
-};
 
 export const getRisingBooks = async () => {
   const data = await fetchRisingBooks();
