+++ src/app/books/[list]/_services/books.ts
@@ -1,11 +1,12 @@
 'use server';
 
-import { getPopularBooks } from '@/app/(home)/_services/books';
+import { getPopularBooks } from '@/services';
 
 interface LoadNewBooksProps {
   list: string;
   nextPage: number;
 }
+
 export const loadNewBooks = async ({ list, nextPage }: LoadNewBooksProps) => {
   if (list === 'popular') {
     const result = await getPopularBooks({ pageNo: nextPage.toString(), pageSize: '10' });
